Refactor(io): move numpy utilities to nputils
